add date searching
This commit adds the date struct and two functions for dealing with
dates.

dateadd returns the date `days` ahead of `x`. `days` can also be
negative, returning the date `days` behind `x` instead. This function
does not account for leap years. It is not needed and overstretches by 1
days so will not make a difference.

`datewithin` returns if `x` is inclusively within dates `low` and
`high`.