add range searching
This commit extends the use `datewithin` to allow for a range of days
to be specified. This range is one-directional (i.e. include the next 5
days or the previous 5 days) from a given date.

The `week` argument/verb is now shorthand for `-i 7`. Using the `-i`
option with `week` does not do anything.