add file sorting
Files are now sorted using alphasort before being printed. Since file
names are (or should be) in the ISO 8601 date format this will sort them
in order of date.

As a result of using scandir, matchfilter and rangefilter have been
seperated out from the main loop. All of the new global variables are a
result of this, but I do not mind it.

This commit also adds support for options that take no arguments. These
options must start with two dashes ('--'). The `--nosort` option will
disable the sorting added in this commit.